Centralize runtime secret env planning
