+++ runtime-agent-ci/lib/full-run-config.cjs
@@ -23,7 +23,11 @@ const {
 const {
   buildSecretEnvFallbacks,
   buildSecretEnvPlan,
-} = require('./runtime-contracts.cjs');
+  normalizeSecretEnvInput,
+  normalizeSecretEnvMap,
+  secretEnvNamesFromRequirements,
+  validateEnvName,
+} = require('./secret-env-plan.cjs');
 
 function main() {
   writeFullRunConfig(process.env);
@@ -312,48 +316,6 @@ function normalizeStringArray(value) {
     return [value];
   }
   return Array.isArray(value) ? value.filter((entry) => typeof entry === 'string' && entry.length > 0) : [];
-}
-
-function normalizeSecretEnvInput(value) {
-  if (typeof value !== 'string' || value.trim() === '') {
-    return [];
-  }
-  const trimmed = value.trim();
-  const entries = trimmed.startsWith('[') ? JSON.parse(trimmed) : splitCsv(trimmed);
-  if (!Array.isArray(entries)) {
-    throw new Error('secret_env must be a JSON array or comma-separated list');
-  }
-  for (const entry of entries) {
-    validateEnvName(entry, 'secret_env');
-  }
-  return entries;
-}
-
-function normalizeSecretEnvMap(map) {
-  const normalized = {};
-  for (const [target, sources] of Object.entries(map || {})) {
-    validateEnvName(target, 'secret_env_map target');
-    const sourceList = Array.isArray(sources) ? sources : [sources];
-    if (sourceList.length === 0) {
-      throw new Error(`secret_env_map.${target} requires at least one source env name`);
-    }
-    normalized[target] = sourceList.map((source) => validateEnvName(source, `secret_env_map.${target}`));
-  }
-  return normalized;
-}
-
-function secretEnvNamesFromRequirements(requirements) {
-  if (!Array.isArray(requirements)) {
-    return [];
-  }
-  return requirements.map((entry) => entry?.name).filter((name) => typeof name === 'string' && name.length > 0);
-}
-
-function validateEnvName(name, label) {
-  if (typeof name !== 'string' || !/^[A-Za-z_][A-Za-z0-9_]*$/.test(name)) {
-    throw new Error(`${label} entries must be valid environment variable names`);
-  }
-  return name;
 }
 
 function uniqueStrings(values) {
