+++ runtime-agent-ci/provider-adapters.js
@@ -8,4 +8,5 @@ Object.assign(module.exports, require('./lib/runtime-workflow-inputs.cjs'));
 Object.assign(module.exports, require('./lib/preview-materialization'));
 Object.assign(module.exports, require('./lib/agent-task-outcome-normalizer'));
 Object.assign(module.exports, require('./lib/runtime-contracts.cjs'));
+Object.assign(module.exports, require('./lib/secret-env-plan.cjs'));
 Object.assign(module.exports, require('./lib/full-run-config.cjs'));
